apps/agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0

# LangGraph + LangChain
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-text-splitters>=0.2.0

# Google AI
google-genai>=0.5.0

# Supabase
supabase>=2.4.0

# Ingestión
pypdf>=4.0.0
pdfplumber>=0.11.0
crawl4ai>=0.3.0
tree-sitter>=0.21.0

# HTTP + utils
httpx>=0.27.0
tenacity>=8.3.0
python-dotenv>=1.0.0

# Dev
pytest>=8.0.0
pytest-asyncio>=0.23.0
